requirements.txt
Flask
Gunicorn

opencv-python
numpy
camera
simpleaudio
Pillow
wave
math

#encoder only
soundfile
pyaudio
io
subprocess